--- a/src/backend/distributed/sql/citus--11.0-3--11.1-1.sql
+++ b/src/backend/distributed/sql/citus--11.0-3--11.1-1.sql
@@ -69,3 +69,4 @@ DROP FUNCTION pg_catalog.get_all_active_transactions(OUT datid oid, OUT process_
 #include "udfs/get_all_active_transactions/11.1-1.sql"
 #include "udfs/citus_split_shard_by_split_points/11.1-1.sql"
 #include "udfs/worker_split_copy/11.1-1.sql"
+#include "udfs/worker_copy_table_to_node/11.1-1.sql"