--- a/src/backend/distributed/sql/udfs/worker_copy_table_to_node/11.1-1.sql
+++ b/src/backend/distributed/sql/udfs/worker_copy_table_to_node/11.1-1.sql
@@ -1,0 +1,8 @@
+CREATE OR REPLACE FUNCTION pg_catalog.worker_copy_table_to_node(
+    source_table regclass,
+    target_node_id integer)
+RETURNS void
+LANGUAGE C STRICT
+AS 'MODULE_PATHNAME', $$worker_copy_table_to_node$$;
+COMMENT ON FUNCTION pg_catalog.worker_copy_table_to_node(regclass, integer)
+    IS 'Perform copy of a shard';