… src/test/regress/expected/tableam.out
… src/test/regress/multi_schedule
… src/test/regress/operations_schedule
… src/test/regress/sql/failure_offline_move_shard_placement.sql
… src/test/regress/sql/shard_rebalancer.sql
… src/test/regress/sql/tableam.sql
… src/test/regress/sql/worker_copy_table_to_node.sql